Numpy 2 Support (#1928)
* initial fix for numpy 2

* allow numpy>=2

* add mdtraj._version into docstring test ignore list

* allow numpy>=2

* pin numpy>=2

* force numpy2 in conda test env

* separate rng from global instance for test_hdf5.py

* fix test_lammpstrj.py

* force all quantities to use float64 or int64

* pin to numpy2, drop ambertools from testing

* two equal signs for pip

* force 2.*

* allow 2.* for conda test env as well

* Revert "allow 2.* for conda test env as well"

This reverts commit 81404205b015fea659f94b52d0362377455e4750.

* another stab at numpy 2 support

* remove numpy2 NEP 50 warnings

* Revert "remove numpy2 NEP 50 warnings"

This reverts commit fe5475685df2c568b4827ea54f1f47bc0cfc1e69.

* test numpy 1 & 2 in CI

* allow different numpy versions in setup.py

* turn weak_and_warn on only on numpy2

* Revert "another stab at numpy 2 support"

This reverts commit 9c7546e6ccd33e0a3f961fda5dee99abf964f969.

* Revert "force all quantities to use float64 or int64"

This reverts commit 6cf803ccd0703a258475205d99ccd85092c27165.

* modify alignment scalars to np.float64

* second round of fixes

* fix dssp tests

* separate each pdb into its own test, allow 'P' to ' ' assignments

* fix test_hbonds dssp test, remove np.core deprecation warning

* replace running mkdssp with reference test files

* mv dssp files into separate folder

* fix dssp test for 2EQQ.pdb

* forgot to rename file for test_hbonds

* try to fix rdf test?

* speed up the test a little bit

* make sure all values are close, bar the orders of magnitude big one

* all fixed now, including thermodynamics

* change value in reference file

* trim down warnings

* be more explicit about dependencies

* support numpy2 and numpy1 in setup/pyproject

* don't be greedy with casting

* make CI description for numpy more verbose

* remove the NEP50 warnings

* allow LAMMPSTrajectory to store lengths, xyz and angles in np.float64

* roll back RDF calculation changes

* add new RDF reference data and mark test

* remove white space

---------

+++ mdtraj/geometry/sasa.py
@@ -276,7 +276,7 @@ def shrake_rupley(
             dtype=np.int32,
         )
         if not np.all(
-            np.unique(atom_mapping) == np.arange(1 + np.max(atom_mapping)),
+            np.unique(atom_mapping) == np.arange(np.int64(1) + np.max(atom_mapping)),
         ):
             raise ValueError(
                 "residues must have contiguous integer indices " "starting from zero",
